+++ scrapy_zyte_api/utils.py
@@ -16,6 +16,7 @@ _SCRAPY_2_1_0 = Version("2.1.0")
 _SCRAPY_2_4_0 = Version("2.4.0")
 _SCRAPY_2_5_0 = Version("2.5.0")
 _SCRAPY_2_6_0 = Version("2.6.0")
+_SCRAPY_2_12_0 = Version("2.12.0")
 
 # Need to install an asyncio reactor before download handler imports to work
 # around:
@@ -24,6 +25,7 @@ _SCRAPY_2_6_0 = Version("2.6.0")
 # https://github.com/scrapy/scrapy/commit/e4bdd1cb958b7d89b86ea66f0af1cec2d91a6d44
 _NEEDS_EARLY_REACTOR = _SCRAPY_2_4_0 <= _SCRAPY_VERSION < _SCRAPY_2_6_0
 
+_AUTOTHROTTLE_DONT_ADJUST_DELAY_SUPPORT = _SCRAPY_VERSION >= _SCRAPY_2_12_0
 _DOWNLOAD_NEEDS_SPIDER = _SCRAPY_VERSION < _SCRAPY_2_6_0
 _RAW_CLASS_SETTING_SUPPORT = _SCRAPY_VERSION >= _SCRAPY_2_4_0
 _REQUEST_ERROR_HAS_QUERY = _PYTHON_ZYTE_API_VERSION >= _PYTHON_ZYTE_API_0_5_2
@@ -45,3 +47,13 @@ except ImportError:  # Scrapy < 2.12
         objcls: type[T], crawler: Crawler, /, *args: Any, **kwargs: Any
     ) -> T:
         return create_instance(objcls, None, crawler, *args, **kwargs)
+
+
+try:
+    import scrapy_poet  # noqa: F401
+except ImportError:
+    _POET_ADDON_SUPPORT = False
+else:
+    _SCRAPY_POET_VERSION = Version(version("scrapy-poet"))
+    _SCRAPY_POET_0_26_0 = Version("0.26.0")
+    _POET_ADDON_SUPPORT = _SCRAPY_POET_VERSION >= _SCRAPY_POET_0_26_0