+++ tests/test_providers.py
@@ -1,4 +1,5 @@
 from collections import defaultdict
+from copy import deepcopy
 from typing import Annotated
 
 import pytest
@@ -10,7 +11,6 @@ from pytest_twisted import ensureDeferred
 from scrapy import Request, Spider
 from scrapy_poet import DummyResponse
 from scrapy_poet.utils.testing import HtmlResource, crawl_single_item
-from scrapy_poet.utils.testing import create_scrapy_settings as _create_scrapy_settings
 from twisted.internet import reactor
 from twisted.web.client import Agent, readBody
 from web_poet import (
@@ -43,6 +43,7 @@ from scrapy_zyte_api import (
     actions,
     custom_attrs,
 )
+from scrapy_zyte_api._params import _EXTRACT_KEYS
 from scrapy_zyte_api.handler import ScrapyZyteAPIDownloadHandler
 from scrapy_zyte_api.providers import _AUTO_PAGES, _ITEM_KEYWORDS, ZyteApiProvider
 
@@ -50,16 +51,6 @@ from . import DEFAULT_AUTOMAP_PARAMS, SETTINGS
 from .mockserver import get_ephemeral_port
 
 PROVIDER_PARAMS = {"geolocation": "IE"}
-
-
-def create_scrapy_settings():
-    settings = _create_scrapy_settings()
-    for setting, value in SETTINGS.items():
-        if setting.endswith("_MIDDLEWARES") and settings[setting]:
-            settings[setting].update(value)
-        else:
-            settings[setting] = value
-    return settings
 
 
 @attrs.define
@@ -108,7 +99,7 @@ class ZyteAPIProviderMetaSpider(Spider):
 
 @ensureDeferred
 async def test_provider(mockserver):
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     item, url, _ = await crawl_single_item(ZyteAPISpider, HtmlResource, settings)
@@ -155,7 +146,7 @@ async def test_itemprovider_requests_direct_dependencies(fresh_mockserver):
     port = get_ephemeral_port()
     handle_urls(f"{fresh_mockserver.host}:{port}")(MyPage)
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = fresh_mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 1100}
     item, url, _ = await crawl_single_item(
@@ -182,7 +173,7 @@ async def test_itemprovider_requests_indirect_dependencies(fresh_mockserver):
     port = get_ephemeral_port()
     handle_urls(f"{fresh_mockserver.host}:{port}")(MyPage)
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = fresh_mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 1100}
     item, url, _ = await crawl_single_item(
@@ -210,7 +201,7 @@ async def test_itemprovider_requests_indirect_dependencies_workaround(fresh_mock
     port = get_ephemeral_port()
     handle_urls(f"{fresh_mockserver.host}:{port}")(MyPage)
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = fresh_mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 1}
     item, url, _ = await crawl_single_item(
@@ -228,7 +219,7 @@ async def test_itemprovider_requests_indirect_dependencies_workaround(fresh_mock
 
 @ensureDeferred
 async def test_provider_params_setting(mockserver):
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_PROVIDER_PARAMS"] = PROVIDER_PARAMS
@@ -239,7 +230,7 @@ async def test_provider_params_setting(mockserver):
 
 @ensureDeferred
 async def test_provider_params_meta(mockserver):
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     _, _, crawler = await crawl_single_item(
@@ -251,7 +242,7 @@ async def test_provider_params_meta(mockserver):
 
 @ensureDeferred
 async def test_provider_params_remove_unused_options(mockserver):
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_PROVIDER_PARAMS"] = {
@@ -281,7 +272,7 @@ async def test_provider_extractfrom(mockserver):
                 "product2": page.product,
             }
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
 
@@ -311,7 +302,7 @@ async def test_provider_extractfrom_double(mockserver, caplog):
                 "product": page.product,
             }
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
 
@@ -332,7 +323,7 @@ async def test_provider_extractfrom_override(mockserver):
                 "product": page.product,
             }
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_PROVIDER_PARAMS"] = {
@@ -365,7 +356,7 @@ async def test_provider_geolocation(mockserver):
                 "product": page.product,
             }
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
 
@@ -384,7 +375,7 @@ async def test_provider_geolocation_unannotated(mockserver, caplog):
         def parse_(self, response: DummyResponse, page: GeoProductPage):  # type: ignore[override]
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
 
@@ -422,7 +413,7 @@ async def test_provider_custom_attrs(mockserver, annotation):
                 "custom_attrs": page.custom_attrs,
             }
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
 
@@ -465,7 +456,7 @@ async def test_provider_custom_attrs_values(mockserver):
                 "custom_attrs": page.custom_attrs,
             }
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
 
@@ -500,7 +491,7 @@ class RecordingHandler(ScrapyZyteAPIDownloadHandler):
 
 
 def provider_settings(server):
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = server.urljoin("/")
     settings["ZYTE_API_TRANSPARENT_MODE"] = True
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 1100}
@@ -1117,7 +1108,7 @@ async def test_provider_actions(mockserver, caplog):
                 "action_results": page.actions,
             }
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
 
@@ -1139,6 +1130,10 @@ async def test_provider_actions(mockserver, caplog):
     )
 
 
+def test_item_keywords():
+    assert set(_EXTRACT_KEYS) == set(_ITEM_KEYWORDS.values())
+
+
 def test_auto_pages_set():
     assert set(_ITEM_KEYWORDS) == {get_item_cls(cls) for cls in _AUTO_PAGES}  # type: ignore[call-overload]
 
@@ -1155,7 +1150,7 @@ async def test_auto_field_stats_not_enabled(mockserver):
         def parse(self, response: DummyResponse, product: Product):
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     _, _, crawler = await crawl_single_item(TestSpider, HtmlResource, settings)
@@ -1210,7 +1205,7 @@ async def test_auto_field_stats_no_override(mockserver):
         def parse(self, response: DummyResponse, product: Product):
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["STATS_CLASS"] = OnlyOnceStatsCollector
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_AUTO_FIELD_STATS"] = True
@@ -1259,7 +1254,7 @@ async def test_auto_field_stats_partial_override(mockserver):
         def parse(self, response: DummyResponse, product: Product):
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_AUTO_FIELD_STATS"] = True
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
@@ -1411,7 +1406,7 @@ async def test_auto_field_stats_full_override(mockserver):
         def parse(self, response: DummyResponse, product: Product):
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_AUTO_FIELD_STATS"] = True
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
@@ -1448,7 +1443,7 @@ async def test_auto_field_stats_callback_override(mockserver):
             product.name = "foo"
             yield product
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_AUTO_FIELD_STATS"] = True
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
@@ -1499,7 +1494,7 @@ async def test_auto_field_stats_item_page_override(mockserver):
         def parse(self, response: DummyResponse, page: MyProductPage):
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_AUTO_FIELD_STATS"] = True
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
@@ -1565,7 +1560,7 @@ async def test_auto_field_stats_alt_page_override(mockserver):
         def parse(self, response: DummyResponse, page: AltProductPage):
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_AUTO_FIELD_STATS"] = True
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
@@ -1616,7 +1611,7 @@ async def test_auto_field_stats_non_auto_override(mockserver):
         def parse(self, response: DummyResponse, product: Product):
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_AUTO_FIELD_STATS"] = True
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
@@ -1661,7 +1656,7 @@ async def test_auto_field_stats_auto_field_decorator(mockserver):
         def parse(self, response: DummyResponse, product: Product):
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_AUTO_FIELD_STATS"] = True
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
@@ -1707,7 +1702,7 @@ async def test_auto_field_stats_auto_field_meta(mockserver):
         def parse(self, response: DummyResponse, product: Product):
             pass
 
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     settings["ZYTE_API_AUTO_FIELD_STATS"] = True
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
@@ -1750,7 +1745,7 @@ class ZyteAPIMultipleSpider(Spider):
 
 @ensureDeferred
 async def test_multiple_types(mockserver):
-    settings = create_scrapy_settings()
+    settings = deepcopy(SETTINGS)
     settings["ZYTE_API_URL"] = mockserver.urljoin("/")
     settings["SCRAPY_POET_PROVIDERS"] = {ZyteApiProvider: 0}
     item, url, _ = await crawl_single_item(