+++ tests/test_sessions.py
@@ -1,7 +1,7 @@
 from collections import deque
 from copy import copy, deepcopy
 from math import floor
-from typing import Any, Dict, Tuple, Union
+from typing import Any, Dict, Optional, Tuple, Union
 from unittest.mock import patch
 
 import pytest
@@ -19,6 +19,7 @@ from scrapy_zyte_api import (
     SESSION_DEFAULT_RETRY_POLICY,
     LocationSessionConfig,
     SessionConfig,
+    get_request_session_id,
     is_session_init_request,
     session_config,
 )
@@ -770,6 +771,54 @@ async def test_max_bad_inits_per_pool(global_setting, pool_setting, value, mocks
     ),
 )
 @ensureDeferred
+async def test_max_check_failures(setting, value, mockserver):
+    retry_times = 2
+    settings = {
+        "RETRY_TIMES": retry_times,
+        "ZYTE_API_RETRY_POLICY": "scrapy_zyte_api.SESSION_DEFAULT_RETRY_POLICY",
+        "ZYTE_API_SESSION_CHECKER": "tests.test_sessions.FalseUseChecker",
+        "ZYTE_API_SESSION_ENABLED": True,
+        "ZYTE_API_SESSION_PARAMS": {"url": "https://example.com"},
+        "ZYTE_API_SESSION_POOL_SIZE": 1,
+        "ZYTE_API_URL": mockserver.urljoin("/"),
+    }
+    if setting is not None:
+        settings["ZYTE_API_SESSION_MAX_CHECK_FAILURES"] = setting
+
+    class TestSpider(Spider):
+        name = "test"
+        start_urls = ["https://example.com"]
+
+        def parse(self, response):
+            pass
+
+    crawler = await get_crawler(settings, spider_cls=TestSpider, setup_engine=False)
+    await crawler.crawl()
+
+    session_stats = {
+        k: v
+        for k, v in crawler.stats.get_stats().items()
+        if k.startswith("scrapy-zyte-api/sessions")
+    }
+    assert session_stats == {
+        "scrapy-zyte-api/sessions/pools/example.com/init/check-passed": floor(
+            (retry_times + 1) / value
+        )
+        + 1,
+        "scrapy-zyte-api/sessions/pools/example.com/use/check-failed": retry_times + 1,
+    }
+
+
+@pytest.mark.parametrize(
+    ("setting", "value"),
+    (
+        (None, 1),
+        (0, 1),
+        (1, 1),
+        (2, 2),
+    ),
+)
+@ensureDeferred
 async def test_max_errors(setting, value, mockserver):
     retry_times = 2
     settings = {
@@ -1211,10 +1260,11 @@ async def test_addon(manual_settings, addon_settings):
             "ZYTE_API_TRANSPARENT_MODE": True,
             "ZYTE_API_SESSION_ENABLED": True,
             **manual_settings,
-        }
+        },
+        poet=False,
     )
     addon_crawler = await get_crawler(
-        {"ZYTE_API_SESSION_ENABLED": True, **addon_settings}, use_addon=True
+        {"ZYTE_API_SESSION_ENABLED": True, **addon_settings}, use_addon=True, poet=False
     )
     assert serialize_settings(crawler.settings) == serialize_settings(
         addon_crawler.settings
@@ -2079,6 +2129,123 @@ async def test_session_config_no_web_poet(mockserver):
         @session_config(["example.com"])
         class CustomSessionConfig(SessionConfig):
             pass
+
+
+@ensureDeferred
+async def test_session_config_process_request_change_request(mockserver):
+    pytest.importorskip("web_poet")
+
+    @session_config("example.com")
+    class CustomSessionConfig(SessionConfig):
+
+        def __init__(self, crawler):
+            super().__init__(crawler)
+            self.session_data = {}
+
+        def check(self, response: Response, request: Request) -> bool:
+            if is_session_init_request(request):
+                session_id = get_request_session_id(request)
+                self.session_data[session_id] = {"foo": "bar"}
+            return super().check(response, request)
+
+        def process_request(self, request: Request) -> Optional[Request]:
+            session_id = get_request_session_id(request)
+            foo = self.session_data[session_id]["foo"]
+            request.headers["foo"] = foo
+
+    settings = {
+        "RETRY_TIMES": 0,
+        "ZYTE_API_URL": mockserver.urljoin("/"),
+        "ZYTE_API_SESSION_ENABLED": True,
+        "ZYTE_API_SESSION_MAX_BAD_INITS": 1,
+    }
+    request_headers = []
+
+    class TestSpider(Spider):
+        name = "test"
+        start_urls = ["https://example.com"]
+
+        def parse(self, response):
+            request_headers.append(response.request.headers["foo"])
+
+    crawler = await get_crawler(settings, spider_cls=TestSpider, setup_engine=False)
+    await crawler.crawl()
+
+    assert request_headers == [b"bar"]
+
+    session_stats = {
+        k: v
+        for k, v in crawler.stats.get_stats().items()
+        if k.startswith("scrapy-zyte-api/sessions")
+    }
+    assert session_stats == {
+        "scrapy-zyte-api/sessions/pools/example.com/init/check-passed": 1,
+        "scrapy-zyte-api/sessions/pools/example.com/use/check-passed": 1,
+    }
+
+    # Clean up the session config registry, and check it, otherwise we could
+    # affect other tests.
+
+    session_config_registry.__init__()  # type: ignore[misc]
+
+
+@ensureDeferred
+async def test_session_config_process_request_new_request(mockserver):
+    pytest.importorskip("web_poet")
+
+    @session_config("example.com")
+    class CustomSessionConfig(SessionConfig):
+
+        def __init__(self, crawler):
+            super().__init__(crawler)
+            self.session_data = {}
+
+        def check(self, response: Response, request: Request) -> bool:
+            if is_session_init_request(request):
+                session_id = get_request_session_id(request)
+                self.session_data[session_id] = {"foo": "bar"}
+            return super().check(response, request)
+
+        def process_request(self, request: Request) -> Optional[Request]:
+            session_id = get_request_session_id(request)
+            foo = self.session_data[session_id]["foo"]
+            new_url = request.url.rstrip("/") + f"/{foo}"
+            return request.replace(url=new_url)
+
+    settings = {
+        "RETRY_TIMES": 0,
+        "ZYTE_API_URL": mockserver.urljoin("/"),
+        "ZYTE_API_SESSION_ENABLED": True,
+        "ZYTE_API_SESSION_MAX_BAD_INITS": 1,
+    }
+    output_urls = []
+
+    class TestSpider(Spider):
+        name = "test"
+        start_urls = ["https://example.com"]
+
+        def parse(self, response):
+            output_urls.append(response.url)
+
+    crawler = await get_crawler(settings, spider_cls=TestSpider, setup_engine=False)
+    await crawler.crawl()
+
+    assert output_urls == ["https://example.com/bar"]
+
+    session_stats = {
+        k: v
+        for k, v in crawler.stats.get_stats().items()
+        if k.startswith("scrapy-zyte-api/sessions")
+    }
+    assert session_stats == {
+        "scrapy-zyte-api/sessions/pools/example.com/init/check-passed": 1,
+        "scrapy-zyte-api/sessions/pools/example.com/use/check-passed": 1,
+    }
+
+    # Clean up the session config registry, and check it, otherwise we could
+    # affect other tests.
+
+    session_config_registry.__init__()  # type: ignore[misc]
 
 
 @ensureDeferred